+++ R/data-converters.R
@@ -272,7 +272,9 @@ enw_incidence_to_linelist <- function(obs, reference_date = "reference_date",
 #' and be complete. See [enw_complete_dates()] for more information. If
 #' NA values are present in the `confirm` column then these will be set to
 #' zero before aggregation this may not be desirable if this missingness
-#' is meaningful.
+#' is meaningful. Before aggregation, dates will be completed up to the
+#' last reference date to ensure all reference dates have the required
+#' report dates for the specified timestep.
 #'
 #' @param min_reference_date The minimum reference date to start the
 #' aggregation from. Note that the timestep will start from the minimum
@@ -306,7 +308,7 @@ enw_aggregate_cumulative <- function(
     required_cols = c("confirm", by), forbidden_cols = ".group",
     dates = TRUE, copy = copy
   )
-
+  obs <- enw_complete_dates(obs, by = by, timestep = "day")
   obs <- enw_assign_group(obs, by = by)
   check_timestep_by_date(obs, timestep = "day", exact = TRUE)
 
@@ -364,7 +366,8 @@ enw_aggregate_cumulative <- function(
     agg_obs_na_ref <- agg_obs_na_ref[report_date_mod == 0]
     agg_obs <- rbind(agg_obs_na_ref, agg_obs, fill = TRUE)
   }
-
+  # Add in any missing new confirm counts
+  agg_obs <- enw_add_incidence(agg_obs, by = by, copy = FALSE)
   # Drop internal processing columns
   agg_obs[,
    c("reference_date_mod", "report_date_mod", ".group") := NULL